moves graphql schema and resolvers out into their own directories
placeholder data still present.
