+++ api/resolvers/resolvers.js
@@ -1,0 +1,12 @@
+const resolvers = {
+  Query: {
+    greet: () => {
+      return 'Hello from GraphQl side';
+    },
+    users: () => {
+      return fetchData();
+    },
+  },
+};
+
+module.exports = resolvers;
